+++ packages/examples/src/testing-utils/getSchema.ts
@@ -1,0 +1,4 @@
+import { buildASTSchema } from "graphql";
+
+// eslint-disable-next-line @typescript-eslint/no-var-requires
+export const getSchema = () => buildASTSchema(require("../schema.graphql"));