+++ packages/nova-react/src/graphql/hooks.test.tsx
@@ -53,7 +53,8 @@ describe(useLazyLoadQuery, () => {
       const { data } = useLazyLoadQuery<{
         response: string;
         variables: Record<string, unknown>;
-      }>(query, {});
+        context?: Record<string, unknown>;
+      }>(query, {}, { context: { callerInfo: "subject-with-query" } });
       return <span>{data}</span>;
     };
 
@@ -63,7 +64,11 @@ describe(useLazyLoadQuery, () => {
       </NovaGraphQLProvider>,
     );
 
-    expect(graphql.useLazyLoadQuery).toHaveBeenCalledWith(query, {}, undefined);
+    expect(graphql.useLazyLoadQuery).toHaveBeenCalledWith(
+      query,
+      {},
+      { context: { callerInfo: "subject-with-query" } }
+    );
     expect(screen.getByText("some-data")).toBeDefined();
   });
 });
@@ -289,6 +294,7 @@ describe(useSubscription, () => {
     const expectedArgument = {
       subscription,
       variables: { someVar: "some-value" },
+      context: { callerInfo: "subject-with-subscription" },
       onNext,
       onError,
     };