+++ packages/nova-react/src/graphql/hooks.ts
@@ -53,12 +53,14 @@ import type {
  * @param variables Object containing the variable values to fetch the query. These variables need to match GraphQL
  *                  variables declared inside the query.
  * @param options Options passed on to the underlying implementation.
+ * @param options.context The query context to pass along the apollo link chain. Should be avoided when possible as
+ *                        it will not be compatible with Relay APIs.
  * @returns An object with either an error, the result data, or neither while loading.
  */
 export function useLazyLoadQuery<TQuery extends OperationType>(
   query: GraphQLTaggedNode,
   variables: TQuery["variables"],
-  options?: { fetchPolicy: "cache-first" },
+  options?: { fetchPolicy?: "cache-first"; context?: TQuery["context"] }
 ): { error?: Error; data?: TQuery["response"] } {
   const graphql = useNovaGraphQL();
   invariant(
@@ -207,6 +209,10 @@ interface GraphQLSubscriptionConfig<
   subscription: GraphQLTaggedNode;
   variables: TSubscriptionPayload["variables"];
   /**
+   * Should be avoided when possible as it will not be compatible with Relay APIs.
+   */
+  context?: TSubscriptionPayload["context"];
+  /**
    * Should response be nullable?
    */
   onNext?: (response: TSubscriptionPayload["response"]) => void;
@@ -226,6 +232,10 @@ export function useSubscription<TSubscriptionPayload extends OperationType>(
 
 interface MutationCommitterOptions<TMutationPayload extends OperationType> {
   variables: TMutationPayload["variables"];
+  /**
+   * Should be avoided when possible as it will not be compatible with Relay APIs.
+   */
+  context?: TMutationPayload["context"];
   optimisticResponse?: Partial<TMutationPayload["response"]> | null;
 }