+++ scripts/config/jest.config.js
@@ -3,4 +3,8 @@ module.exports = {
   rootDir: process.cwd(),
   roots: ["<rootDir>/src"],
   testPathIgnorePatterns: ["node_modules", "__generated__"],
+  testEnvironment: "jsdom",
+  transform: {
+    "\\.(gql|graphql)$": "@graphql-tools/jest-transform",
+  },
 };